feat: Move middleware_manager to core/managers/
- Move src/fapilog/_internal/middleware_manager.py to src/fapilog/core/managers/
- Update import statements in:
  - src/fapilog/_internal/component_factory.py
  - src/fapilog/container.py
  - tests/test_middleware_manager_integration.py
  - tests/test_component_factory.py
  - tests/test_middleware_manager.py
- Fix relative imports within middleware_manager.py
- Fix dynamic import path in middleware_manager.py
- Update all patch paths in test files (30+ references updated)

✅ MiddlewareManager successfully moved to new location (4/12 files moved)
